+++ packages/sdk/src/sdk.test.js
@@ -24,25 +24,27 @@ describe("build", () => {
     assert.deepEqual(txProposer.sequenceNum, 123)
   })
 
-  it("accepts an async function as a param", async () => {
-    const identity = {
-      asParam: v => v,
-      asInjection: v => v,
-    }
+  // TODO: Uncomment once interaction params have resolve field 
 
-    const ix = await resolve(await build([
-      transaction``,
-      params([ async () => {
-        return {
-          key: "my_param", value: 1, xform: identity
-        }
-      } ])
-    ]), [ resolveParams ])
+  // it("accepts an async function as a param", async () => {
+  //   const identity = {
+  //     asParam: v => v,
+  //     asInjection: v => v,
+  //   }
 
-    const ixParams = Object.values(ix.params)
-    const p = ixParams.find(p => p.key === "my_param")
+  //   const ix = await resolve(await build([
+  //     transaction``,
+  //     params([ async () => {
+  //       return {
+  //         key: "my_param", value: 1, xform: identity
+  //       }
+  //     } ])
+  //   ]), [ resolveParams ])
 
-    assert.deepEqual(p.key, "my_param")
-    assert.deepEqual(p.value, 1)
-  })
+  //   const ixParams = Object.values(ix.params)
+  //   const p = ixParams.find(p => p.key === "my_param")
+
+  //   assert.deepEqual(p.key, "my_param")
+  //   assert.deepEqual(p.value, 1)
+  // })
 })